requirements.txt
absl-py==0.8.1
astor==0.8.0
cycler==0.10.0
DateTime==4.3
Django==2.2.6
django-bootstrap-modal-forms==1.4.4
django-bootstrap3==11.1.0
et-xmlfile==1.0.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
jdcal==1.4.1
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mysqlclient==1.4.4
numpy==1.17.2
openpyxl==3.0.0
opt-einsum==3.1.0
pandas==0.25.1
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
zope.interface==4.6.0
